+++ scripts/util.py
@@ -14,6 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import inspect
+import json
 import httplib2
 import importlib.util
 import logging
@@ -77,8 +79,8 @@ if ENV_CONFIG_YAML:
 else:
     CONFIG_FILE = Path(__file__).with_name("config.yaml")
 API_REQ_LIMIT = 2000
+URI_REGEX = r"[a-z]([-a-z0-9]*[a-z0-9])?"
 
-log = logging.getLogger(__name__)
 def_creds, auth_project = google.auth.default()
 Path.mkdirp = partialmethod(Path.mkdir, parents=True, exist_ok=True)
 
@@ -124,6 +126,59 @@ slurmdirs = NSDict(
         )
     }
 )
+
+
+class LogFormatter(logging.Formatter):
+    """adds logging flags to the levelname in log records"""
+
+    def format(self, record):
+        new_fmt = self._fmt
+        flag = getattr(record, "flag", None)
+        if flag is not None:
+            start, level, end = new_fmt.partition("%(levelname)s")
+            if level:
+                new_fmt = f"{start}{level}(%(flag)s){end}"
+        # insert function name if record level is DEBUG
+        if record.levelno < logging.INFO:
+            prefix, msg, suffix = new_fmt.partition("%(message)s")
+            new_fmt = f"{prefix}%(funcName)s: {msg}{suffix}"
+        self._style._fmt = new_fmt
+        return super().format(record)
+
+
+class FlagLogAdapter(logging.LoggerAdapter):
+    """creates log adapters that add a flag to the log record,
+    allowing it to be filtered"""
+
+    def __init__(self, logger, flag, extra=None):
+        if extra is None:
+            extra = {}
+        self.flag = flag
+        super().__init__(logger, extra)
+
+    @property
+    def enabled(self):
+        return cfg.extra_logging_flags.get(self.flag, False)
+
+    def process(self, msg, kwargs):
+        extra = kwargs.setdefault("extra", {})
+        extra.update(self.extra)
+        extra["flag"] = self.flag
+        return msg, kwargs
+
+
+logging.basicConfig(level=logging.INFO, stream=sys.stdout)
+log = logging.getLogger(__name__)
+logging_flags = [
+    "trace_api",
+    "subproc",
+    "hostlists",
+    "subscriptions",
+]
+log_trace_api = FlagLogAdapter(log, "trace_api")
+log_subproc = FlagLogAdapter(log, "subproc")
+log_hostlists = FlagLogAdapter(log, "hostlists")
+log_subscriptions = FlagLogAdapter(log, "subscriptions")
 
 
 def publish_message(project_id, topic_id, message) -> None:
@@ -262,7 +317,8 @@ def subscription_create(subscription_id, project_id=None):
         }
         try:
             subscription = subscriber.create_subscription(request=request)
-            log.info(f"Subscription created: {subscription}")
+            log.info(f"Subscription created: {subscription_path}")
+            log_subscriptions.debug(f"{subscription}")
         except exceptions.AlreadyExists:
             log.info(f"Subscription '{subscription_path}' already exists!")
 
@@ -349,6 +405,12 @@ def load_config_data(config):
         cfg.slurm_bin_dir = slurmdirs.prefix / "bin"
     if not cfg.slurm_control_host:
         cfg.slurm_control_host = f"{cfg.slurm_cluster_name}-controller"
+
+    if not cfg.enable_debug_logging and isinstance(cfg.enable_debug_logging, NSDict):
+        cfg.enable_debug_logging = False
+    cfg.extra_logging_flags = NSDict(
+        {flag: cfg.extra_logging_flags.get(flag, False) for flag in logging_flags}
+    )
     return cfg
 
 
@@ -409,12 +471,26 @@ def save_config(cfg, path):
     Path(path).write_text(yaml.dump(cfg, Dumper=Dumper))
 
 
-def config_root_logger(
-    caller_logger, level="DEBUG", util_level=None, stdout=True, logfile=None
-):
+def filter_logging_flags(record):
+    """logging filter for flags
+    if there are no flags, always pass. If there are flags, only pass if a flag
+    matches an enabled flag in cfg.extra_logging_flags"""
+    flag = getattr(record, "flag", None)
+    if flag is None:
+        return True
+    return cfg.extra_logging_flags.get(flag, False)
+
+
+def owned_file_handler(filename):
+    """create file handler"""
+    if filename is None:
+        return None
+    chown_slurm(filename)
+    return logging.handlers.WatchedFileHandler(filename, delay=True)
+
+
+def config_root_logger(caller_logger, level="DEBUG", stdout=True, logfile=None):
     """configure the root logger, disabling all existing loggers"""
-    if not util_level:
-        util_level = level
     handlers = list(compress(("stdout_handler", "file_handler"), (stdout, logfile)))
 
     config = {
@@ -422,23 +498,31 @@ def config_root_logger(
         "disable_existing_loggers": True,
         "formatters": {
             "standard": {
-                "format": "",
+                "()": LogFormatter,
+                "fmt": "%(levelname)s: %(message)s",
             },
             "stamp": {
-                "format": "%(asctime)s %(name)s %(levelname)s: %(message)s",
+                "()": LogFormatter,
+                "fmt": "%(asctime)s %(levelname)s: %(message)s",
             },
+        },
+        "filters": {
+            "logging_flags": {"()": lambda: filter_logging_flags},
         },
         "handlers": {
             "stdout_handler": {
-                "level": "DEBUG",
+                "level": logging.DEBUG,
                 "formatter": "standard",
                 "class": "logging.StreamHandler",
                 "stream": sys.stdout,
+                "filters": ["logging_flags"],
             },
-        },
-        "loggers": {
-            __name__: {  # enable util.py logging
-                "level": util_level,
+            "file_handler": {
+                "()": owned_file_handler,
+                "level": logging.DEBUG,
+                "formatter": "stamp",
+                "filters": ["logging_flags"],
+                "filename": logfile,
             },
         },
         "root": {
@@ -446,15 +530,11 @@ def config_root_logger(
             "level": level,
         },
     }
-    if logfile:
-        config["handlers"]["file_handler"] = {
-            "level": "DEBUG",
-            "formatter": "stamp",
-            "class": "logging.handlers.WatchedFileHandler",
-            "filename": logfile,
-        }
+    if not logfile:
+        del config["handlers"]["file_handler"]
     logging.config.dictConfig(config)
     loggers = (
+        __name__,
         "resume",
         "suspend",
         "slurmsync",
@@ -463,6 +543,19 @@ def config_root_logger(
     )
     for logger in map(logging.getLogger, loggers):
         logger.disabled = False
+
+
+def log_api_request(request):
+    """log.trace info about a compute API request"""
+    if log_trace_api.enabled:
+        # output the whole request object as pretty yaml
+        # the body is nested json, so load it as well
+        rep = json.loads(request.to_json())
+        if rep.get("body", None) is not None:
+            rep["body"] = json.loads(rep["body"])
+        pretty_req = yaml.safe_dump(rep).rstrip()
+        # label log message with the calling function
+        log_trace_api.debug(f"{inspect.stack()[1].function}:\n{pretty_req}")
 
 
 def handle_exception(exc_type, exc_value, exc_trace):
@@ -484,7 +577,7 @@ def run(
     **kwargs,
 ):
     """Wrapper for subprocess.run() with convenient defaults"""
-    log.debug(f"run: {cmd}")
+    log_subproc.debug(f"run: {cmd}")
     args = cmd if shell else shlex.split(cmd)
     result = subprocess.run(
         args,
@@ -502,7 +595,7 @@ def run(
 def spawn(cmd, quiet=False, shell=False, **kwargs):
     """nonblocking spawn of subprocess"""
     if not quiet:
-        log.debug(f"spawn: {cmd}")
+        log_subproc.debug(f"spawn: {cmd}")
     args = cmd if shell else shlex.split(cmd)
     return subprocess.Popen(args, shell=shell, **kwargs)
 
@@ -661,10 +754,9 @@ def to_hostlist(nodenames):
     tmp_file = tempfile.NamedTemporaryFile(mode="w+t", delete=False)
     tmp_file.writelines("\n".join(sorted(nodenames, key=natural_sort)))
     tmp_file.close()
-    log.debug("tmp_file = {}".format(tmp_file.name))
 
     hostlist = run(f"{lkp.scontrol} show hostlist {tmp_file.name}").stdout.rstrip()
-    log.debug("hostlist = {}".format(hostlist))
+    log_hostlists.debug(f"hostlist({len(nodenames)}): {hostlist}".format(hostlist))
     os.remove(tmp_file.name)
     return hostlist
 
@@ -676,6 +768,7 @@ def to_hostnames(nodelist):
     else:
         hostlist = ",".join(nodelist)
     hostnames = run(f"{lkp.scontrol} show hostnames {hostlist}").stdout.splitlines()
+    log_hostlists.debug(f"hostnames({len(hostnames)}) from {hostlist}")
     return hostnames
 
 
@@ -807,7 +900,7 @@ def wait_for_operation(operation, project=None, compute=compute):
         if result["status"] == "DONE":
             log_errors = " with errors" if "error" in result else ""
             log.debug(
-                f"operation complete{log_errors}: type: {result['operationType']}, name: {result['name']}"
+                f"operation complete{log_errors}: type={result['operationType']}, name={result['name']}"
             )
             return result
 
@@ -880,15 +973,18 @@ def get_filtered_operations(
     return operations
 
 
-def get_insert_operations(bulk_operations, project=None, compute=compute):
-    """get all group operations from list of bulk operations"""
+def get_insert_operations(group_ids, flt=None, project=None, compute=compute):
+    """get all insert operations from a list of operationGroupId"""
     if project is None:
         project = lkp.project
-    flt = " OR ".join(
-        f"(operationGroupId={op['operationGroupId']})" for op in bulk_operations
-    )
-    predicate = f"AND ({flt})" if flt else ""
-    return get_filtered_operations(f"(operationType=insert) {predicate}")
+    if isinstance(group_ids, str):
+        group_ids = group_ids.split(",")
+    filters = [
+        "operationType=insert",
+        flt,
+        " OR ".join(f"(operationGroupId={id})" for id in group_ids),
+    ]
+    return get_filtered_operations(" AND ".join(f"({f})" for f in filters if f))
 
 
 class Dumper(yaml.SafeDumper):